refactor: remove index fiiles from components
