+++ site/src/components/EmptyState/EmptyState.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from "@testing-library/react"
 import React from "react"
 import { render } from "../../testHelpers"
-import { EmptyState } from "./index"
+import { EmptyState } from "./EmptyState"
 
 describe("EmptyState", () => {
   it("renders (smoke test)", async () => {
