+++ site/src/components/ErrorSummary/ErrorSummary.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentMeta, Story } from "@storybook/react"
 import React from "react"
-import { ErrorSummary, ErrorSummaryProps } from "."
+import { ErrorSummary, ErrorSummaryProps } from "./ErrorSummary"
 
 export default {
   title: "components/ErrorSummary",
