+++ site/src/components/ErrorSummary/ErrorSummary.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react"
 import React from "react"
-import { ErrorSummary } from "./index"
+import { ErrorSummary } from "./ErrorSummary"
 
 describe("ErrorSummary", () => {
   it("renders", async () => {
