+++ site/src/components/Navbar/UserDropdown/UserDropdown.stories.tsx
@@ -1,7 +1,7 @@
 import Box from "@material-ui/core/Box"
 import { Story } from "@storybook/react"
 import React from "react"
-import { UserDropdown, UserDropdownProps } from "."
+import { UserDropdown, UserDropdownProps } from "./UsersDropdown"
 
 export default {
   title: "Page/UserDropdown",
