--- site/src/components/Page/index.tsx
@@ -1,2 +1,0 @@
-export * from "./Footer"
-export * from "./RequireAuth"
