--- site/src/components/SignIn/index.tsx
@@ -1,2 +1,0 @@
-export * from "./CliAuthToken"
-export * from "./SignInForm"
