+++ site/src/components/Snackbar/utils.ts
@@ -1,12 +1,6 @@
 import { dispatchCustomEvent } from "../../util/events"
 
 ///////////////////////////////////////////////////////////////////////////////
-// Notification Component
-///////////////////////////////////////////////////////////////////////////////
-
-export { GlobalSnackbar } from "./GlobalSnackbar"
-
-///////////////////////////////////////////////////////////////////////////////
 // Notification Types
 ///////////////////////////////////////////////////////////////////////////////
