+++ site/src/components/TabPanel/TabPanel.stories.tsx
@@ -1,6 +1,6 @@
 import { Story } from "@storybook/react"
 import React from "react"
-import { TabPanel, TabPanelProps } from "."
+import { TabPanel, TabPanelProps } from "./TabPanel"
 
 export default {
   title: "TabPanel/TabPanel",
