+++ site/src/components/Table/Cells/UserCell.tsx
@@ -3,7 +3,7 @@ import Link from "@material-ui/core/Link"
 import { makeStyles } from "@material-ui/core/styles"
 import Typography from "@material-ui/core/Typography"
 import React from "react"
-import { UserAvatar, UserAvatarProps } from "../../User"
+import { UserAvatarProps, UserAvatar } from "../../User/UserAvatar"
 
 export interface UserCellProps {
   Avatar: UserAvatarProps
