--- site/src/components/Table/index.tsx
@@ -1,1 +1,0 @@
-export * from "./Table"
