--- site/src/components/User/index.ts
@@ -1,2 +1,0 @@
-export * from "./UserAvatar"
-export * from "./UserProfileCard"
