--- site/src/components/Workspace/index.ts
@@ -1,1 +1,0 @@
-export * from "./Workspace"
