--- site/src/components/index.tsx
@@ -1,2 +1,0 @@
-export * from "./EmptyState"
-export * from "./Page"
