+++ site/src/forms/CreateTemplateForm.tsx
@@ -5,14 +5,12 @@ import React from "react"
 import * as Yup from "yup"
 import { CreateTemplateRequest, Organization, Provisioner, Template } from "../api/types"
 import { LoadingButton } from "../components/Button/LoadingButton"
-import {
-  DropdownItem,
-  FormCloseButton,
-  FormDropdownField,
-  FormSection,
-  FormTextField,
-  FormTitle,
-} from "../components/Form"
+import { FormCloseButton } from "../components/Form/FormCloseButton"
+import { DropdownItem, FormDropdownField } from "../components/Form/FormDropdownField"
+import { FormSection } from "../components/Form/FormSection"
+import { FormTextField } from "../components/Form/FormTextField"
+import { FormTitle } from "../components/Form/FormTitle"
+
 
 export interface CreateTemplateFormProps {
   provisioners: Provisioner[]
