+++ site/src/pages/cli-auth.tsx
@@ -3,7 +3,7 @@ import { useActor } from "@xstate/react"
 import React, { useContext, useEffect, useState } from "react"
 import { getApiKey } from "../api"
 import { FullScreenLoader } from "../components/Loader/FullScreenLoader"
-import { CliAuthToken } from "../components/SignIn"
+import { CliAuthToken } from "../components/SignIn/CliAuthToken"
 import { XServiceContext } from "../xServices/StateContext"
 
 export const CliAuthenticationPage: React.FC = () => {
