+++ site/src/pages/templates/[organization]/[template]/create.tsx
@@ -4,7 +4,7 @@ import { useNavigate, useParams } from "react-router-dom"
 import useSWR from "swr"
 import * as API from "../../../../api"
 import * as Types from "../../../../api/types"
-import { ErrorSummary } from "../../../../components/ErrorSummary"
+import { ErrorSummary } from "../../../../components/ErrorSummary/ErrorSummary"
 import { FullScreenLoader } from "../../../../components/Loader/FullScreenLoader"
 import { CreateWorkspaceForm } from "../../../../forms/CreateWorkspaceForm"
 import { unsafeSWRArgument } from "../../../../util"
