+++ site/src/pages/templates/[organization]/[template]/index.tsx
@@ -4,11 +4,11 @@ import React from "react"
 import { Link, useNavigate, useParams } from "react-router-dom"
 import useSWR from "swr"
 import { Organization, Template, Workspace } from "../../../../api/types"
-import { EmptyState } from "../../../../components/EmptyState"
-import { ErrorSummary } from "../../../../components/ErrorSummary"
-import { Header } from "../../../../components/Header"
+import { EmptyState } from "../../../../components/EmptyState/EmptyState"
+import { ErrorSummary } from "../../../../components/ErrorSummary/ErrorSummary"
+import { Header } from "../../../../components/Header/Header"
 import { FullScreenLoader } from "../../../../components/Loader/FullScreenLoader"
-import { Column, Table } from "../../../../components/Table"
+import { Column, Table } from "../../../../components/Table/Table"
 import { unsafeSWRArgument } from "../../../../util"
 import { firstOrItem } from "../../../../util/array"
