+++ site/src/testHelpers/index.tsx
@@ -3,7 +3,7 @@ import { render as wrappedRender, RenderResult } from "@testing-library/react"
 import { createMemoryHistory } from "history"
 import React from "react"
 import { MemoryRouter, Route, Routes, unstable_HistoryRouter as HistoryRouter } from "react-router-dom"
-import { RequireAuth } from "../components"
+import { RequireAuth } from "../components/Page/RequireAuth"
 import { dark } from "../theme"
 import { XServiceProvider } from "../xServices/StateContext"
 import { MockUser } from "./entities"
