+++ site/src/xServices/auth/authXService.ts
@@ -1,7 +1,7 @@
 import { assign, createMachine } from "xstate"
 import * as API from "../../api"
 import * as Types from "../../api/types"
-import { displaySuccess } from "../../components/Snackbar"
+import { displaySuccess } from "../../components/Snackbar/utils"
 
 export const Language = {
   successProfileUpdate: "Updated preferences.",
